requirements.txt
requests>=2.31.0
pandas>=1.5.0
scikit-learn>=1.0.0
urllib3>=2.0.0
numpy>=1.23.0
SQLAlchemy>=2.0.0
solana>=0.30.0
anchorpy>=0.15.0
base58>=2.1.1
python-telegram-bot>=20.0
web3>=6.11.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
hummingbot==1.11.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
coverage==7.3.2
binance-connector>=3.0.0
kucoin-python>=1.0.0
cryptography>=40.0.0
pycryptodome>=3.18.0
psycopg2-binary>=2.9.9
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
asyncpg>=0.28.0
redis>=5.0.0
cachetools>=5.3.0
